# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /care/consultations:
    post:
      operationId: createConsultation
      tags: [care]
      summary: Book a doctor consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationInput"
      responses:
        "201":
          description: Consultation booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareRequest"
        "400":
          description: Invalid consultation input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /care/lab-bookings:
    post:
      operationId: createLabBooking
      tags: [care]
      summary: Book a lab test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabBookingInput"
      responses:
        "201":
          description: Lab booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareRequest"
        "400":
          description: Invalid lab booking input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /care/medicine-orders:
    post:
      operationId: createMedicineOrder
      tags: [care]
      summary: Create a medicine delivery request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicineOrderInput"
      responses:
        "201":
          description: Medicine order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareRequest"
        "400":
          description: Invalid medicine order input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    CareRequest:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [consultation, lab, medicine]
        itemId:
          type: string
        title:
          type: string
        patientName:
          type: string
        phone:
          type: string
        notes:
          type: string
        address:
          type: string
        mode:
          type: string
        dateSlot:
          type: string
        status:
          type: string
        amount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, type, itemId, title, patientName, phone, notes, address, mode, dateSlot, status, amount, createdAt]
...
    CreateConsultationInput:
      type: object
      properties:
        doctorId:
          type: string
          minLength: 1
        patientName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 10
        concern:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [video, clinic]
        dateSlot:
          type: string
          minLength: 1
      required: [doctorId, patientName, phone, concern, mode, dateSlot]
    CreateLabBookingInput:
      type: object
      properties:
        testId:
          type: string
          minLength: 1
        patientName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 10
        address:
          type: string
          minLength: 1
        dateSlot:
          type: string
          minLength: 1
      required: [testId, patientName, phone, address, dateSlot]
    CreateMedicineOrderInput:
      type: object
      properties:
        medicineId:
          type: string
          minLength: 1
        patientName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 10
        address:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
      required: [medicineId, patientName, phone, address, quantity]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error